requirements/base.txt
django
django-auditlog
django-braces
django-handyhelpers
